HBP-87. Move elephant module under tvb_multiscale.
